2023/01/11
1. controller 작업 완료

2. 2023/01/12 즉 내일은 이미지, 동영상 업로드 관련 작업.

+++ src/main/java/Baetube_backEnd/service/history/HistoryDeleteService.java
@@ -2,16 +2,24 @@ package Baetube_backEnd.service.history;
 
 import org.springframework.beans.factory.annotation.Autowired;
 
-import Baetube_backEnd.dto.DeleteHistoryRequest;
+import Baetube_backEnd.dto.History;
+import Baetube_backEnd.exception.NullHistoryException;
 import Baetube_backEnd.mapper.HistoryMapper;
 
 public class HistoryDeleteService
 {
 	@Autowired
 	private HistoryMapper historyMapper;
-	
-	public boolean deleteHistory(DeleteHistoryRequest request)
+
+	public boolean deleteHistory(History request)
 	{
+		History history = historyMapper.select(request.getUserId(), request.getVideoId());
+		
+		if(history == null)
+		{
+			throw new NullHistoryException();
+		}
+		
 		historyMapper.delete(request.getUserId(), request.getVideoId());
 		
 		return true;
